+++ src/syngen/VERSION
@@ -1,1 +1,1 @@
-0.10.7rc6
+0.10.7rc9
